# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 健身房後端 API — 最終作業規格書
  version: 1.0.0
  description: |
    ## 這本規格書怎麼用

    左側依**里程碑**分組（M1 → M6 就是建議的施工順序）。每支 API 都有
    request / response 範例：先看範例、再看欄位說明，需要登入的 API
    按右上角 **Authorize** 貼上 login 拿到的 token，就能直接 **Try it out** 試打你的後端。

    ## 全域回應格式（每一支都適用）

    | 情況 | HTTP 狀態碼 | Body |
    |---|---|---|
    | 成功 | 2xx | `{ "status": "success", "data": ... }` |
    | 可預期的失敗（驗證錯、找不到、沒權限） | 4xx | `{ "status": "failed", "message": "..." }` |
    | 你的程式壞掉 | 5xx | （不在驗收範圍，但請努力不要發生） |

    - **驗收只看 2xx / 4xx + status 欄位**：文件裡寫的 200/201 是參考值，回 200 或 201 都算對
    - 需要登入的 API 一律收 `Authorization: Bearer <token>`
    - JWT payload 必須含 `{ id, role, exp }`（前端會解開來用，缺了會壞）
    - 所有 id 都是 uuid 字串
    - 日期時間欄位用 UTC ISO 8601 字串（例 `2026-08-20T10:00:00Z`）

    ## ⚠️ 四句固定錯誤訊息文字（全文件最重要的一段）

    以下四句錯誤訊息，前端會拿來判斷要顯示哪個畫面，必須一字不差：

    1. `已經報名過此課程`
    2. `已無可使用堂數`
    3. `已達最大參加人數，無法參加`
    4. `請先登入`

    其他錯誤訊息文字隨你寫（只要 status 是 failed）。
servers:
  - url: http://localhost:8080
    description: 你的後端（本機開發）
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用 POST /api/users/login 拿到的 token
paths:
...
  /api/coaches/skill:
    get:
      tags:
        - "M1 種資料：技能與方案"
      summary: 取得教練技能列表
      description: |-
        回傳全部技能，不需登入。前端有三個地方共用這支：管理者的技能管理頁、教練個人資料頁、開課表單的技能下拉選單。
        ⚠️ 路由順序地雷：這條路徑必須排在 `/api/coaches/{coachId}`（教練詳情）**之前**。兩條路徑長得都像「/api/coaches/後面接一段文字」，如果教練詳情那條先註冊，`skill` 這個字會被當成 coachId 吃掉——這支 API 就永遠打不到。掛路由時請把這條放前面。
      responses:
        "200":
          description: 成功取得技能列表（還沒有任何技能時回空陣列 `[]`，不是錯誤）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: 技能 id（uuid 字串）
                        name:
                          type: string
                          description: 技能名稱
              example:
                status: success
                data:
                  - id: "2a4b6c8d-1e2f-4a5b-9c8d-0e1f2a3b4c5d"
                    name: "重訓"
                  - id: "5f6e7d8c-9b0a-4c3d-8e7f-6a5b4c3d2e1f"
                    name: "瑜伽"
    post:
      tags:
        - "M1 種資料：技能與方案"
      summary: 新增教練技能
      description: |-
        新增一筆技能，不需登入（課程簡化設計：這是管理用的「種資料」端點，正式產品會加權限控管）。
        技能名稱不可重複。
        ⚠️ 成功狀態碼是 **200，不是 201**（答案版實際行為）。驗收只看「成功 2xx ＋ status 欄位」，所以你回 201 也算過，但文件以 200 為準。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: 技能名稱，必填、不可為空字串或全空白
            example:
              name: "重訓"
      responses:
        "200":
          description: 新增成功，回傳完整的技能資料（含後端產生的 id 與建立時間）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: 後端產生的 uuid
                      name:
                        type: string
                      createdAt:
                        type: string
                        description: 建立時間（ISO 8601 字串）
              example:
                status: success
                data:
                  id: "2a4b6c8d-1e2f-4a5b-9c8d-0e1f2a3b4c5d"
                  name: "重訓"
                  createdAt: "2026-06-10T08:00:00.000Z"
        "400":
          description: 觸發條件：沒給 name、name 不是字串、或 name 是空字串／全空白
          content:
            application/json:
              example:
                status: failed
                message: "欄位未填寫正確"
        "409":
          description: 觸發條件：name 與既有技能重複
          content:
            application/json:
              example:
                status: failed
                message: "資料重複"
...
  /api/credit-package:
    get:
      tags:
        - "M1 種資料：技能與方案"
      summary: 取得購買方案列表
      description: |-
        回傳全部堂數方案，不需登入。前端的「健身方案」頁直接吃這份資料。
        ⚠️ `price` 的型別：答案版回的是**數字字串**（例如 `"1400.00"`），你回**數字**（例如 `1400`）也可以——前端兩種都處理得了，不構成驗收差異。文件範例照答案版寫字串。
      responses:
        "200":
          description: 成功取得方案列表（還沒有任何方案時回空陣列 `[]`，不是錯誤）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: 方案 id（uuid 字串）
                        name:
                          type: string
                          description: 方案名稱
                        credit_amount:
                          type: integer
                          description: 此方案包含的堂數
                        price:
                          description: 價格。數字或數字字串皆可（答案版回字串 "1400.00"）
                          oneOf:
                            - type: number
                            - type: string
              example:
                status: success
                data:
                  - id: "0d2c1b3a-4e5f-4a6b-8c7d-9e0f1a2b3c4d"
                    name: "7 堂組合包方案"
                    credit_amount: 7
                    price: "1400.00"
                  - id: "1e3d2c4b-5f6a-4b7c-9d8e-0f1a2b3c4d5e"
                    name: "14 堂組合包方案"
                    credit_amount: 14
                    price: "2520.00"
    post:
      tags:
        - "M1 種資料：技能與方案"
      summary: 新增購買方案
      description: |-
        新增一筆堂數方案，不需登入（課程簡化的管理端點；前端頁面沒有呼叫它，但它是把「健身方案」資料種進系統的唯一入口——不先用這支種資料，方案頁會永遠空白、購買流程也無從驗收）。
        方案名稱不可重複。`credit_amount` 與 `price` 都必須是 **0 以上的整數（數字型別）**——price 請送數字，不要送字串，送字串會被擋 400。
        ⚠️ 成功狀態碼是 **200，不是 201**（答案版實際行為）。驗收只看「成功 2xx ＋ status 欄位」，回 201 也算過。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - credit_amount
                - price
              properties:
                name:
                  type: string
                  description: 方案名稱，必填、不可為空字串或全空白、不可與既有方案重複
                credit_amount:
                  type: integer
                  description: 方案包含的堂數，必填、0 以上的整數
                price:
                  type: integer
                  description: 方案價格，必填、0 以上的整數（要送數字型別，不能送字串）
            example:
              name: "7 堂組合包方案"
              credit_amount: 7
              price: 1400
      responses:
        "200":
          description: 新增成功，回傳完整的方案資料（含後端產生的 id 與建立時間）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: 後端產生的 uuid
                      name:
                        type: string
                      credit_amount:
                        type: integer
                      price:
                        type: number
                      createdAt:
                        type: string
                        description: 建立時間（ISO 8601 字串）
              example:
                status: success
                data:
                  id: "0d2c1b3a-4e5f-4a6b-8c7d-9e0f1a2b3c4d"
                  name: "7 堂組合包方案"
                  credit_amount: 7
                  price: 1400
                  createdAt: "2026-06-10T08:00:00.000Z"
        "400":
          description: 觸發條件：任一欄位沒給；name 不是字串或為空；credit_amount 或 price 不是數字、是負數、或帶小數
          content:
            application/json:
              example:
                status: failed
                message: "欄位未填寫正確"
        "409":
          description: 觸發條件：name 與既有方案重複
          content:
            application/json:
              example:
                status: failed
                message: "資料重複"
...
  /api/users/credit-package:
    get:
      tags:
        - M5 購買與報名
      summary: 取得本人的購買方案紀錄
      description: |
        回傳目前登入者買過的所有堂數方案紀錄。data 直接是「陣列」（不是物件包陣列），
        沒買過任何方案時回空陣列 []。

        行為備註：
        - 每筆形狀固定：{ name（方案名稱）, purchased_credits（買到的堂數）,
          price_paid（付了多少錢，數字）, purchase_at（購買時間）}。
        - ⚠️ price_paid 回「數字」型別（例如 1400，不是字串 "1400.00"）。
        - 排序：依 purchase_at 新到舊（最新買的排最前面）。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功取得購買紀錄（沒買過回空陣列）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: 方案名稱
                          example: 14 堂組合包方案
                        purchased_credits:
                          type: integer
                          description: 這筆購買取得的堂數
                          example: 14
                        price_paid:
                          type: integer
                          description: 實際支付金額（數字型別）
                          example: 1400
                        purchase_at:
                          type: string
                          description: 購買時間（UTC ISO 8601）
                          example: 2026-05-10T03:20:00.000Z
              example:
                status: success
                data:
                  - name: 14 堂組合包方案
                    purchased_credits: 14
                    price_paid: 1400
                    purchase_at: "2026-05-10T03:20:00.000Z"
                  - name: 7 堂組合包方案
                    purchased_credits: 7
                    price_paid: 1400
                    purchase_at: "2026-04-01T09:00:00.000Z"
        "401":
          description: |
            token 驗證失敗。下方是建議回應範例；驗收重點是 4xx + status: failed。只有「請先登入」是固定錯誤訊息文字，必須完全相同。
            - 沒帶 Authorization header、或格式不是 Bearer <token>：「請先登入」
            - token 已過期：「Token 已過期」
            - token 無效（內容不對、或查無此使用者）：「無效的 token」
            ⚠️「請先登入」是四句固定錯誤訊息文字之一，一個字都不能改。
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                未帶token:
                  summary: 沒帶 Authorization header 或缺 token
                  value:
                    status: failed
                    message: 請先登入
                token過期:
                  summary: token 已過期
                  value:
                    status: failed
                    message: Token 已過期
                token無效:
                  summary: token 無效或查無此使用者
                  value:
                    status: failed
                    message: 無效的 token
...
  /api/admin/coaches:
    get:
      tags:
        - M3 教練後台
      summary: 取得教練本人的後台資料（含技能清單）
      description: |
        教練登入後查看自己的教練資料，通常用來把資料塞進「編輯教練資料」的表單。需要登入，而且登入的人必須已經是教練。

        行為備註：
        - 回傳的 `id` 是**教練資料的 id**，不是使用者 id，兩個是不同的東西，不要拿去當 userId 用。
        - `skill_ids` 是這位教練目前綁定的技能 id 清單（uuid 字串陣列）；還沒設定過技能時會回空陣列 `[]`。
        - 剛升級、還沒填過大頭貼的教練，`profile_image_url` 可能是 null。⚠️ 把這份資料塞進表單後直接按儲存會吃 400，原因看 PUT 那支的說明。

        ⚠️ **固定錯誤訊息文字**：沒帶 token 時回的「請先登入」是前端會直接拿來判斷畫面的四句固定錯誤訊息文字之一，**一字不能差**（不能寫成「請登入」「尚未登入」等變體），否則前端成品會壞掉。

        文件寫的狀態碼是答案版實際行為，供參考；**驗收只看「成功 2xx / 失敗 4xx」加上回應的 status 欄位**。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功取得教練本人資料
          content:
            application/json:
              example:
                status: success
                data:
                  id: "a1c5e8f2-3b7d-4e9a-8c6f-2d4b9e1a7c3f"
                  experience_years: 3
                  description: "瑜伽與皮拉提斯雙修，擅長帶初學者"
                  profile_image_url: "https://example.com/avatar.png"
                  skill_ids:
                    - "7e2f9a4b-1c8d-4f6e-b3a5-9d7c2e4f8a1b"
                    - "3c6e1f8a-5b2d-4a9c-8e7f-1b4d6a9c3e5f"
        "401":
          description: |
            登入或身分檢查失敗，依情境回不同訊息：
            - 沒帶 Authorization header 或沒帶 token →「請先登入」（⚠️ 固定錯誤訊息文字，一字不差）
            - token 已過期 →「Token 已過期」
            - token 無效或查無此使用者 →「無效的 token」
            - 有登入但還不是教練 →「使用者尚未成為教練」（注意：這裡回的是 401，不是 403）
          content:
            application/json:
              examples:
                請先登入:
                  summary: 沒帶 token（⚠️ 訊息為固定錯誤訊息文字）
                  value:
                    status: failed
                    message: "請先登入"
                Token 已過期:
                  summary: token 過期
                  value:
                    status: failed
                    message: "Token 已過期"
                無效的 token:
                  summary: token 無效或查無使用者
                  value:
                    status: failed
                    message: "無效的 token"
                使用者尚未成為教練:
                  summary: 已登入但 role 不是 COACH
                  value:
                    status: failed
                    message: "使用者尚未成為教練"
    put:
      tags:
        - M3 教練後台
      summary: 更新教練本人的後台資料（含整批更換技能）
      description: |
        教練更新自己的年資、介紹、大頭貼與技能清單。需要登入，而且登入的人必須已經是教練。成功後回傳更新完的同形狀資料，前端可以直接拿來覆蓋表單。

        ⚠️ **profile_image_url 規則跟「升級教練」那支不一樣**，這是最容易踩的坑，對照表如下：

        | 規則 | 升級教練 POST /api/admin/coaches/{userId} | 更新教練資料 PUT /api/admin/coaches |
        |---|---|---|
        | 是否必填 | 選填（可不給、可空字串） | **必填**（不給或空字串直接 400） |
        | 格式要求 | 有值時必須以 https 開頭 | 必須以 https 開頭 |

        ⚠️ 連鎖地雷：升級時沒填大頭貼的教練，GET 拿到的 `profile_image_url` 是 null；前端把它原樣塞進表單、使用者沒補就按儲存，這支會回 400「欄位未填寫正確」。做前端串接時記得引導使用者先補上 https 開頭的網址。

        行為備註：
        - `skill_ids` 是**整批覆蓋**：你送出的清單會完全取代原本的技能清單（不是「新增」也不是「合併」）。想保留原本的技能，就要連舊的一起送。
        - `skill_ids` 必須是**非空陣列**，且每個元素都是技能的 id（uuid 字串）；空陣列回 400「欄位未填寫正確」。
        - 文件寫的狀態碼是答案版實際行為，供參考；**驗收只看「成功 2xx / 失敗 4xx」加上回應的 status 欄位**。
        - 沒帶 token 回的「請先登入」是固定錯誤訊息文字（同 GET 的說明），一字不能差。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - experience_years
                - description
                - profile_image_url
                - skill_ids
              properties:
                experience_years:
                  type: integer
                  minimum: 0
                  description: 教練年資（數字、不能是負數或小數）
                description:
                  type: string
                  description: 教練介紹（不能是空字串）
                profile_image_url:
                  type: string
                  description: 大頭貼網址。這支是必填，且必須以 https 開頭（⚠️ 跟升級教練那支的選填規則不同）
                skill_ids:
                  type: array
                  minItems: 1
                  items:
                    type: string
                    format: uuid
                  description: 技能 id 清單（整批覆蓋原有技能，不能是空陣列）
              example:
                experience_years: 5
                description: "重訓與體態雕塑，協助學員建立規律訓練習慣"
                profile_image_url: "https://example.com/new-avatar.png"
                skill_ids:
                  - "7e2f9a4b-1c8d-4f6e-b3a5-9d7c2e4f8a1b"
                  - "3c6e1f8a-5b2d-4a9c-8e7f-1b4d6a9c3e5f"
      responses:
        "200":
          description: 更新成功，回傳更新後的教練資料（形狀與 GET 相同，可直接覆蓋表單）
          content:
            application/json:
              example:
                status: success
                data:
                  id: "a1c5e8f2-3b7d-4e9a-8c6f-2d4b9e1a7c3f"
                  experience_years: 5
                  description: "重訓與體態雕塑，協助學員建立規律訓練習慣"
                  profile_image_url: "https://example.com/new-avatar.png"
                  skill_ids:
                    - "7e2f9a4b-1c8d-4f6e-b3a5-9d7c2e4f8a1b"
                    - "3c6e1f8a-5b2d-4a9c-8e7f-1b4d6a9c3e5f"
        "400":
          description: |
            欄位缺漏或格式不對，一律回「欄位未填寫正確」。會觸發的情境：
            - experience_years 不是 0 以上的整數
            - description 沒給或是空字串
            - profile_image_url 沒給、是空字串、或不是 https 開頭（⚠️ 這支是必填）
            - skill_ids 沒給、不是陣列、是空陣列、或元素不是有效的 id 字串
          content:
            application/json:
              example:
                status: failed
                message: "欄位未填寫正確"
        "401":
          description: |
            登入或身分檢查失敗，依情境回不同訊息：
            - 沒帶 Authorization header 或沒帶 token →「請先登入」（⚠️ 固定錯誤訊息文字，一字不差）
            - token 已過期 →「Token 已過期」
            - token 無效或查無此使用者 →「無效的 token」
            - 有登入但還不是教練 →「使用者尚未成為教練」（注意：這裡回的是 401，不是 403）
          content:
            application/json:
              examples:
                請先登入:
                  summary: 沒帶 token（⚠️ 訊息為固定錯誤訊息文字）
                  value:
                    status: failed
                    message: "請先登入"
                Token 已過期:
                  summary: token 過期
                  value:
                    status: failed
                    message: "Token 已過期"
                無效的 token:
                  summary: token 無效或查無使用者
                  value:
                    status: failed
                    message: "無效的 token"
                使用者尚未成為教練:
                  summary: 已登入但 role 不是 COACH
                  value:
                    status: failed
                    message: "使用者尚未成為教練"
  # ===== g4-admin-courses =====
  /api/admin/coaches/courses:
    get:
      tags:
        - M3 教練後台
      summary: 取得教練本人開設的全部課程列表
      description: |
        教練後台「課程管理」頁的列表資料。需要登入，且登入者必須已經是教練（role 為 COACH）。

        行為備註：
        - 只會回「自己開的課」，依登入者的 token 判斷，不會看到別的教練的課程。
        - ⚠️ `status` 是三態中文字串（`尚未開始`／`進行中`／`已結束`），由伺服器**收到請求的當下時間**即時計算，不是存起來的欄位。同一堂課今天查是「尚未開始」，開課後再查就變「進行中」。
        - ⚠️ `participants` 是該課程「未取消」的報名數——已取消的報名不計。
        - 還沒開過任何課時回 `data: []`（空陣列），這是成功回應，不是錯誤。
        - 驗收口徑：文件標的狀態碼數字供參考，驗收只看 2xx／4xx 與 body 的 `status` 欄位。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功取得課程列表（沒開過課時 data 為空陣列）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    description: 教練本人的課程陣列，無課程時為空陣列
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: 課程 id（uuid）
                          example: 5e3b1f0a-9c1d-4e2f-8a7b-1c2d3e4f5a6b
                        name:
                          type: string
                          example: 重訓基礎入門
                        status:
                          type: string
                          description: 三態中文字串，用當下時間即時計算
                          enum:
                            - 尚未開始
                            - 進行中
                            - 已結束
                          example: 尚未開始
                        start_at:
                          type: string
                          description: 課程開始時間（UTC ISO 8601）
                          example: "2026-08-20T10:00:00.000Z"
                        end_at:
                          type: string
                          description: 課程結束時間（UTC ISO 8601）
                          example: "2026-08-20T12:00:00.000Z"
                        max_participants:
                          type: integer
                          example: 10
                        meeting_url:
                          type: string
                          example: "https://meet.example.com/abc-defg-hij"
                        participants:
                          type: integer
                          description: 未取消的報名數。⚠️ 答案版有報名時會回數字字串（如 "2"）、無報名時回數字 0，驗收兩種型別都收，建議統一回數字
                          example: 2
              example:
                status: success
                data:
                  - id: 5e3b1f0a-9c1d-4e2f-8a7b-1c2d3e4f5a6b
                    name: 重訓基礎入門
                    status: 尚未開始
                    start_at: "2026-08-20T10:00:00.000Z"
                    end_at: "2026-08-20T12:00:00.000Z"
                    max_participants: 10
                    meeting_url: "https://meet.example.com/abc-defg-hij"
                    participants: 2
        "401":
          description: |
            未登入、token 有問題、或登入者不是教練時觸發。
            ⚠️ 「請先登入」是前端會拿來判斷畫面的固定錯誤訊息文字，一字不差。
            觸發條件對照：沒帶 Authorization header 或格式不對 →「請先登入」；token 過期 →「Token 已過期」；token 無效或查無使用者 →「無效的 token」；登入了但不是教練 →「使用者尚未成為教練」（注意是 401 不是 403）。
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                未帶token:
                  summary: 沒帶 Authorization header
                  value:
                    status: failed
                    message: 請先登入
                token過期:
                  summary: token 已過期
                  value:
                    status: failed
                    message: Token 已過期
                token無效:
                  summary: token 無效或查無使用者
                  value:
                    status: failed
                    message: 無效的 token
                不是教練:
                  summary: 登入者 role 不是 COACH
                  value:
                    status: failed
                    message: 使用者尚未成為教練
    post:
      tags:
        - M3 教練後台
      summary: 教練開設新課程
      description: |
        教練後台「新增課程」modal 送出時呼叫。需要登入，且登入者必須已經是教練。

        行為備註：
        - ⚠️ 全部欄位都是必填，缺一個、或是空字串，就回「欄位未填寫正確」。
        - ⚠️ `meeting_url` 必須以 `https` 開頭，否則一樣回「欄位未填寫正確」。
        - ⚠️ `max_participants` 必須是「數字型別」的 0 以上整數——送字串 `"10"` 會被擋下來。
        - `start_at`／`end_at` 收 UTC ISO 8601 字串（例 `2026-08-20T10:00:00Z`），原樣存入、原樣吐回。
        - 開課的教練就是 token 裡的本人，body 不用（也不能）指定教練。
        - 成功回 201（不是 200），data.course 是完整課程物件。驗收口徑：狀態碼數字供參考，驗收只看 2xx／4xx 與 body 的 `status` 欄位。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - skill_id
                - name
                - description
                - start_at
                - end_at
                - max_participants
                - meeting_url
              properties:
                skill_id:
                  type: string
                  description: 技能 id（uuid），來自 GET /api/coaches/skill 的下拉選單
                  example: 1c8da31a-5fb2-4f2b-9d3e-6a7b8c9d0e1f
                name:
                  type: string
                  example: 重訓基礎入門
                description:
                  type: string
                  example: 從零開始的重訓課，教你正確使用器材與姿勢。
                start_at:
                  type: string
                  description: 課程開始時間，UTC ISO 8601 字串
                  example: "2026-08-20T10:00:00Z"
                end_at:
                  type: string
                  description: 課程結束時間，UTC ISO 8601 字串
                  example: "2026-08-20T12:00:00Z"
                max_participants:
                  type: integer
                  description: ⚠️ 必須是數字型別的 0 以上整數，字串會被擋
                  example: 10
                meeting_url:
                  type: string
                  description: ⚠️ 必須以 https 開頭
                  example: "https://meet.example.com/abc-defg-hij"
            example:
              skill_id: 1c8da31a-5fb2-4f2b-9d3e-6a7b8c9d0e1f
              name: 重訓基礎入門
              description: 從零開始的重訓課，教你正確使用器材與姿勢。
              start_at: "2026-08-20T10:00:00Z"
              end_at: "2026-08-20T12:00:00Z"
              max_participants: 10
              meeting_url: "https://meet.example.com/abc-defg-hij"
      responses:
        "201":
          description: 開課成功，回完整課程物件（包在 data.course 裡）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      course:
                        type: object
                        description: 完整課程物件
                        properties:
                          id:
                            type: string
                            example: 5e3b1f0a-9c1d-4e2f-8a7b-1c2d3e4f5a6b
                          user_id:
                            type: string
                            description: 開課教練的使用者 id（uuid）
                            example: 9a8b7c6d-5e4f-3a2b-1c0d-e1f2a3b4c5d6
                          skill_id:
                            type: string
                            example: 1c8da31a-5fb2-4f2b-9d3e-6a7b8c9d0e1f
                          name:
                            type: string
                            example: 重訓基礎入門
                          description:
                            type: string
                            example: 從零開始的重訓課，教你正確使用器材與姿勢。
                          start_at:
                            type: string
                            example: "2026-08-20T10:00:00.000Z"
                          end_at:
                            type: string
                            example: "2026-08-20T12:00:00.000Z"
                          max_participants:
                            type: integer
                            example: 10
                          meeting_url:
                            type: string
                            example: "https://meet.example.com/abc-defg-hij"
                          created_at:
                            type: string
                            example: "2026-08-01T03:00:00.000Z"
                          updated_at:
                            type: string
                            example: "2026-08-01T03:00:00.000Z"
              example:
                status: success
                data:
                  course:
                    id: 5e3b1f0a-9c1d-4e2f-8a7b-1c2d3e4f5a6b
                    user_id: 9a8b7c6d-5e4f-3a2b-1c0d-e1f2a3b4c5d6
                    skill_id: 1c8da31a-5fb2-4f2b-9d3e-6a7b8c9d0e1f
                    name: 重訓基礎入門
                    description: 從零開始的重訓課，教你正確使用器材與姿勢。
                    start_at: "2026-08-20T10:00:00.000Z"
                    end_at: "2026-08-20T12:00:00.000Z"
                    max_participants: 10
                    meeting_url: "https://meet.example.com/abc-defg-hij"
                    created_at: "2026-08-01T03:00:00.000Z"
                    updated_at: "2026-08-01T03:00:00.000Z"
        "400":
          description: |
            欄位驗證失敗時觸發：任一欄位缺漏或為空字串、max_participants 不是數字型別的 0 以上整數、meeting_url 不是 https 開頭。
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              example:
                status: failed
                message: 欄位未填寫正確
        "401":
          description: |
            未登入、token 有問題、或登入者不是教練時觸發。
            ⚠️ 「請先登入」是前端會拿來判斷畫面的固定錯誤訊息文字，一字不差。
            觸發條件對照：沒帶 token →「請先登入」；token 過期 →「Token 已過期」；token 無效 →「無效的 token」；不是教練 →「使用者尚未成為教練」。
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                未帶token:
                  summary: 沒帶 Authorization header
                  value:
                    status: failed
                    message: 請先登入
                token過期:
                  summary: token 已過期
                  value:
                    status: failed
                    message: Token 已過期
                token無效:
                  summary: token 無效或查無使用者
                  value:
                    status: failed
                    message: 無效的 token
                不是教練:
                  summary: 登入者 role 不是 COACH
                  value:
                    status: failed
                    message: 使用者尚未成為教練
...